+++ WrapperBundle/Core/Traits/RichTextConvertingEntity.php
@@ -11,14 +11,31 @@ trait RichTextConvertingEntity
 {
     protected $richTextConverter;
 
+    /**
+     * @param Converter $richTextConverter
+     * @return $this
+     */
     public function setRichTextConverter(Converter $richTextConverter)
     {
         $this->richTextConverter = $richTextConverter;
         return $this; // fluent interfaces for setters
     }
 
+    /**
+     * @param string $xml the value of an xml-text field
+     * @return string
+     */
     protected function getHtml($xml)
     {
         return $this->richTextConverter->convert($xml);
+    }
+
+    /**
+     * @param $fieldName the identifier of an xml-text field
+     * @return string
+     */
+    protected function getHtmlForField($fieldName)
+    {
+        return $this->getHtml($this->content()->getFieldValue($fieldName)->xml);
     }
 }