+++ WrapperBundle/Tests/TestEntity.php
@@ -23,4 +23,24 @@ class TestEntity extends BaseEntity
     {
         return parent::getEntityManager();
     }
+
+    public function getHtmlFor($fieldName)
+    {
+        return $this->getHtml($this->content()->getFieldValue($fieldName)->xml);
+    }
+
+    public function getURL($absolute = false)
+    {
+        return $this->getURLAlias($absolute);
+    }
+
+    public function getRelationsFor($fieldName)
+    {
+        return $this->getRelations($fieldName);
+    }
+
+    public function getRelationFor($fieldName)
+    {
+        return $this->getRelation($fieldName);
+    }
 }